+++ .gitignore
@@ -161,4 +161,4 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
-data_collection_pipeline/collected_data/
+artifacts/data/collected_data/
