+++ utils/config.py
@@ -1,0 +1,7 @@
+import yaml
+
+def load_config(path="params.yaml"):
+    with open(path, "r", encoding="utf-8") as f:
+        return yaml.safe_load(f)
+
+CONFIG = load_config()
